requirements.txt
joblib==1.1.0
numpy==1.19.2
Pillow==9.2.0
scikit_learn==1.1.1
tensorboardX==2.5.1
torch==1.8.2+cu111
torchvision==0.9.2+cu111
